Extract the cookie mechanics into a dedicated util module.
Reading `Set-Cookie` headers, deciding which URLs a cookie applies to
and writing a `Cookie` header are all RFC 6265 concerns that have
nothing to do with latchkey's services or credentials. They now live in
cookieUtils, next to the other protocol-level helpers, leaving
cookieCapture.ts as just the login flow: its parameters, its session,
and turning captured cookies into curl arguments.

The module notes where this duplicates the browser: Chromium parses
these headers itself and Playwright exposes the result through
`context.cookies(url)`, which is the alternative if reading responses
directly stops paying for itself.
