+++ src/services/core/cookieCapture.ts
@@ -1,7 +1,8 @@
 /**
  * A generic browser login that works for any cookie-authenticated service:
  * open a URL, watch the `Set-Cookie` headers of the responses that follow, and
- * capture named cookies as the credentials.
+ * capture named cookies as the credentials. The cookie mechanics themselves
+ * live in `cookieUtils`.
  *
  * Nothing here is service-specific — the cookie names come from whoever
  * registered the service, as the parameters of the `cookie-capture` login flow.
@@ -10,6 +11,14 @@
 import type { Response } from 'playwright';
 import { z } from 'zod';
 import { type ApiCredentials, RawCurlCredentials } from '../../apiCredentials/base.js';
+import {
+  cookieScopeKey,
+  doesCookieApplyTo,
+  formatCookieHeaderValue,
+  parseSetCookieHeader,
+  type CookiePair,
+  type ParsedSetCookie,
+} from '../../cookieUtils.js';
 import { Service, SimpleServiceSession } from './base.js';
 
 export const CookieCaptureParamsSchema = z
@@ -26,120 +35,9 @@ export const CookieCaptureParamsSchema = z
 
 export type CookieCaptureParams = z.infer<typeof CookieCaptureParamsSchema>;
 
-/** A captured cookie, reduced to what goes into the header. */
-export interface CookiePair {
-  readonly name: string;
-  readonly value: string;
-}
-
-/**
- * Build the credentials for a set of captured cookies.
- *
- * All pairs go into a single `Cookie` header separated by `'; '`, which is the
- * only portable form: RFC 6265 has the user agent send one `Cookie` header,
- * and servers that see several header lines commonly read just the first.
- */
+/** Store captured cookies the way `latchkey auth set -H` would. */
 export function buildCookieCredentials(cookies: readonly CookiePair[]): ApiCredentials {
-  const header = cookies.map((cookie) => `${cookie.name}=${cookie.value}`).join('; ');
-  return new RawCurlCredentials(['-H', `Cookie: ${header}`]);
-}
-
-/** A `Set-Cookie` header, reduced to what decides whether we want it. */
-export interface ParsedSetCookie {
-  readonly name: string;
-  readonly value: string;
-  /** Effective domain, without the leading dot: the `Domain` attribute if
-   * present, otherwise the host of the response that set the cookie. */
-  readonly domain: string;
-  readonly path: string;
-  /** Whether the header removes the cookie rather than setting it. */
-  readonly isDeletion: boolean;
-}
-
-/**
- * The path a cookie without a `Path` attribute applies to: the directory of the
- * request path (RFC 6265 §5.1.4).
- */
-function defaultCookiePath(responseUrl: URL): string {
-  const path = responseUrl.pathname;
-  if (!path.startsWith('/')) {
-    return '/';
-  }
-  const lastSeparatorIndex = path.lastIndexOf('/');
-  return lastSeparatorIndex === 0 ? '/' : path.slice(0, lastSeparatorIndex);
-}
-
-function isExpired(expiresAttribute: string): boolean {
-  const expiresAt = Date.parse(expiresAttribute);
-  return !Number.isNaN(expiresAt) && expiresAt <= Date.now();
-}
-
-/**
- * Parse one `Set-Cookie` header value. Returns null when it is not a usable
- * cookie definition at all.
- */
-export function parseSetCookieHeader(
-  headerValue: string,
-  responseUrl: URL
-): ParsedSetCookie | null {
-  const [nameValuePair = '', ...attributeParts] = headerValue.split(';');
-  const separatorIndex = nameValuePair.indexOf('=');
-  if (separatorIndex <= 0) {
-    return null;
-  }
-  const name = nameValuePair.slice(0, separatorIndex).trim();
-  const value = nameValuePair.slice(separatorIndex + 1).trim();
-
-  let domain = responseUrl.hostname.toLowerCase();
-  let path = defaultCookiePath(responseUrl);
-  // An empty value is how a cookie is cleared without an explicit expiry.
-  let isDeletion = value === '';
-
-  for (const attributePart of attributeParts) {
-    const attributeSeparatorIndex = attributePart.indexOf('=');
-    const attributeName = (
-      attributeSeparatorIndex === -1
-        ? attributePart
-        : attributePart.slice(0, attributeSeparatorIndex)
-    )
-      .trim()
-      .toLowerCase();
-    const attributeValue =
-      attributeSeparatorIndex === -1 ? '' : attributePart.slice(attributeSeparatorIndex + 1).trim();
-
-    if (attributeName === 'domain' && attributeValue !== '') {
-      domain = attributeValue.replace(/^\./, '').toLowerCase();
-    } else if (attributeName === 'path' && attributeValue.startsWith('/')) {
-      path = attributeValue;
-    } else if (attributeName === 'max-age' && Number(attributeValue) <= 0) {
-      isDeletion = true;
-    } else if (attributeName === 'expires' && isExpired(attributeValue)) {
-      isDeletion = true;
-    }
-  }
-
-  return { name, value, domain, path, isDeletion };
-}
-
-/**
- * Whether a browser would send this cookie to the given URL: the URL's host is
- * the cookie's domain or below it, and its path is under the cookie's path.
- */
-export function doesCookieApplyTo(cookie: ParsedSetCookie, url: URL): boolean {
-  const host = url.hostname.toLowerCase();
-  if (host !== cookie.domain && !host.endsWith(`.${cookie.domain}`)) {
-    return false;
-  }
-  const path = url.pathname === '' ? '/' : url.pathname;
-  return path.startsWith(cookie.path);
-}
-
-/**
- * Identity of a cookie in the jar: the same name can be set more than once
- * under different scopes, and a browser then sends all of them.
- */
-function cookieScopeKey(cookie: ParsedSetCookie): string {
-  return `${cookie.name}\n${cookie.domain}\n${cookie.path}`;
+  return new RawCurlCredentials(['-H', `Cookie: ${formatCookieHeaderValue(cookies)}`]);
 }
 
 /**
